shelf-product-identifier/requirements.txt
ultralytics==8.0.92
numpy==1.24.3
scikit-learn==1.3.0
opencv-python>=4.5.0
Pillow>=8.0.0
transformers>=4.30.0
torch>=1.12.0
torchvision>=0.13.0